Calls to Native must be applied
Fully applied calls are less expensive than qualified calls,
both in terms of space and time.
With elm stuff, we should be able to say "is this a constant?
ok, it's going to be a no-arg function. otherwise it'll be a one-arg function"
Since we can't do any type analysis on Native stuff though,
we would otherwise have to have some other arbitrary restriction.
Like "all native functions have to be wrapped in an empty function"
or "you can't have a native constant".
I think enforcing that native calls are applied is a bit nicer
since it gives you performance benefits as well!